+++ game_object.h
@@ -4,18 +4,18 @@
 
 class GameObject {
  public:
-  GameObject() : x(0.0), y(0.0), width(0) {}
+  GameObject() : x(0), y(0), width(0) {}
   virtual ~GameObject() {}
 
   virtual void Draw() const = 0;
 
-  inline void SetPos(double nx, double ny) {
+  inline void SetPos(int nx, int ny) {
     x = nx;
     y = ny;
   }
 
-  inline double GetY() const { return y; }
-  inline double GetX() const { return x; }
+  inline int GetY() const { return y; }
+  inline int GetX() const { return x; }
 
   inline void SetWidth(uint16_t widthN) { width = widthN; }
   inline uint16_t GetWidth() const { return width; }
